# added token_types

+++ syfertext/doc.py
@@ -282,7 +282,7 @@ class Doc(AbstractObject):
         # Convert to Numpy array.
         # token_vectors = np.array(token_vectors)
         # Convert to torch tensor.
-        token_vectors = torch.stack(tensors=token_vectors)
+        token_vectors = torch.stack(tensors=token_vectors, dtype=torch.float32)
 
         return token_vectors
