+++ tests/local_mode/test_language.py
@@ -3,7 +3,8 @@ import syft as sy
 import torch
 import syfertext
 from pathlib import Path
-import numpy as np
+
+import torch
 
 hook = sy.TorchHook(torch)
 me = hook.local_worker
@@ -40,7 +41,7 @@ def test_vector_valid_token_is_not_zero():
     doc = nlp("banana")
     actual = doc[0].vector
     # changed from numpy to torch zeros tensor
-    zeros = torch.zeros(actual.shape)
+    zeros = torch.zeros(actual.shape, dtype=torch.float32)
 
     # check that at least one cell in actual vector is not zero
     assert (actual != zeros).any() == True
@@ -52,7 +53,7 @@ def test_vector_non_valid_token_is_zero():
     doc = nlp("outofvocabularytoken")
     actual = doc[0].vector
     # changed from numpy to torch zeros tensor
-    zeros = torch.zeros(actual.shape)
+    zeros = torch.zeros(actual.shape, dtype=torch.float32)
 
     # check that all cells in actual vector are zeros
     assert (actual == zeros).all() == True
